Move legacy module to __init__
